update to new evalhyd API
The new API of `evalhyd` is centred around two `evaluate` functions,
one for deterministic evaluation and one for probabilistic evaluation.
The probabilitist side of things is not header-only anymore, and so
additional C++ source files needed to be pointed to in *setup.py*.
The example notebooks are also updated to reflect those changes.