include examples in field-list
since `breathe` systematically puts the field-list for R/C++ API at
the end of the function doc, examples has been added as the last field
in the field-list. So, even though there is no such problem for the
Python API (because it does not need `breathe`, since only required
when linking Doxygen XML), it is better to keep things consistent.